// src/app/types/movie.ts
export interface Genre {
    id: number;
    name: string;
  }
  
  export interface ProductionCompany {
    id: number;
    logo_path?: string ; 
    name: string;
    origin_country: string;
  }
  
  export interface ProductionCountry {
    iso_3166_1: string;
    name: string;
  }
  
  export interface SpokenLanguage {
    english_name: string;
    iso_639_1: string;
    name: string;
  }
  
  export interface Movie {
    adult: boolean;
    backdrop_path: string | null; 
    belongs_to_collection?: {
      backdrop_path: string | null;
      id: number;
      name: string;
      poster_path: string | null;
    };
    budget: number;
    genres: Genre[];
    homepage: string | null; 
    id: number;
    imdb_id: string | null; 
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string | null;
    production_companies: ProductionCompany[];
    production_countries: ProductionCountry[];
    release_date: string;
    revenue: number;
    runtime: number; 
    spoken_languages: SpokenLanguage[];
    status: string;
    tagline: string | null; 
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number; 
  }